refactor: service 인터페이스랑 구현 클래스로 나눔 [#10]
