+++ src/main/java/com/nhnacademy/member_server/controller/MemberController.java
@@ -1,7 +1,7 @@
 package com.nhnacademy.member_server.controller;
 
-import com.nhnacademy.member_server.service.AuthService;
-import com.nhnacademy.member_server.service.MemberService;
+import com.nhnacademy.member_server.service.impl.AuthServiceImpl;
+import com.nhnacademy.member_server.service.impl.MemberServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseCookie;
@@ -13,8 +13,8 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class MemberController {
 
-    private final MemberService memberService; // 회원 상태 관리
-    private final AuthService authService;
+    private final MemberServiceImpl memberServiceImpl; // 회원 상태 관리
+    private final AuthServiceImpl authServiceImpl;
 
     @GetMapping("/my-page")
     public ResponseEntity<String> getMyPage(
@@ -25,8 +25,8 @@ public class MemberController {
 
     @DeleteMapping("/withdraw")
     public ResponseEntity<Void> withdraw(@RequestHeader(name = "X-USER-ID") long loginId) {
-        memberService.withdraw(loginId);
-        authService.logout(loginId);
+        memberServiceImpl.withdraw(loginId);
+        authServiceImpl.logout(loginId);
         ResponseCookie deleteCookie = ResponseCookie.from("refresh-token", "")
                 .path("/")
                 .httpOnly(true)
